+++ getEMagLsFilters.m
@@ -177,6 +177,10 @@ if applyDiffusenessConst
     W_MLS_r = conj(HCorr(:,:,2));
 end
 
+% mamnually set the DC bin
+W_MLS_l(1, :) = abs(W_MLS_l(2, :));
+W_MLS_r(1, :) = abs(W_MLS_r(2, :));
+
 % transform into time domain
 if isreal(Y_Hi_conj)
     W_MLS_l = [W_MLS_l(1:numPosFreqs, :); flipud(conj(W_MLS_l(2:numPosFreqs-1, :)))];