+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Unreleased
+- Update `getEMagLsFilters.m` and `getEMagLs2Filters.m` to manually set the DC bin after calculation
 - Update `getEMagLsFilters.m` with temporary fix leaving out the DC bin
 - Update `getEMagLsFilters.m` to export data for debugging of system-dependent differences in resulting rendering filters
 - Update `getSMAIRMatrix.m` to export data for debugging of system-dependent differences in resulting rendering filters